code review :) for the DisNinjas part
So project looks good and functions properly. Nice job. However, there is some refactoring that we can do. An important concept in coding is repeating yourself as little as possible. While easily done with variables and functions, it also applies to structure. So, what I would challenge you to look at is a way to reduce the amount of if/elif/else that you have handling which image to show. Think about 2 things, 1)how could you dynamically create an image url and 2)How to generalize what you are doing in your if/else logic. Maybe using a different data structure to hold image information to access the images differently. Also, in the future, please try to keep your projects separated and clean the <<<<HEAD and ====== stuff out of your server.py file thanks!